requirements.txt
absl-py
appnope
argon2-cffi
astunparse
async-generator
attrs
backcall
bleach
cachetools
certifi
cffi
chardet
click
cycler
decorator
defusedxml
entrypoints
Flask
gast
google-auth
google-auth-oauthlib
google-pasta
grpcio
h5py
idna
imageio
ipykernel
ipython
ipython-genutils
ipywidgets
itsdangerous
jedi
Jinja2
joblib
jsonschema
jupyter
jupyter-client
jupyter-console
jupyter-core
jupyterlab-pygments
Keras
Keras-Preprocessing
kiwisolver
Markdown
MarkupSafe
matplotlib
mistune
nbclient
nbconvert
nbformat
nest-asyncio
networkx
notebook
numpy
oauthlib
opencv-python
opt-einsum
packaging
pandas
pandocfilters
parso
pexpect
pickleshare
Pillow
prometheus-client
prompt-toolkit
protobuf
ptyprocess
pyasn1
pyasn1-modules
pycparser
Pygments
pyparsing
pyrsistent
python-dateutil
pytz
PyWavelets
PyYAML
pyzmq
qtconsole
QtPy
requests
requests-oauthlib
rsa
scikit-image
scikit-learn
scipy
seaborn
Send2Trash
six
SQLAlchemy
tensorboard
tensorboard-plugin-wit
tensorflow
tensorflow-estimator
termcolor
terminado
testpath
threadpoolctl
tifffile
tornado
traitlets
urllib3
wcwidth
webencodings
Werkzeug
widgetsnbextension
wrapt
